webserver/requirements/local.txt
-r ./base.txt

Werkzeug==0.14.1  # https://github.com/pallets/werkzeug
ipdb==0.11  # https://github.com/gotcha/ipdb
prompt_toolkit==2.0.10 # Required to fix implicit dependency updates in ipdb: https://github.com/ipython/ipython/issues/11962
Sphinx==1.7.5  # https://github.com/sphinx-doc/sphinx
psycopg2==2.7.4 --no-binary psycopg2  # https://github.com/psycopg/psycopg2

# Django
# ------------------------------------------------------------------------------
django-debug-toolbar==1.9.1  # https://github.com/jazzband/django-debug-toolbar
django-extensions==2.0.7  # https://github.com/django-extensions/django-extensions
django-debug-toolbar-request-history==0.0.7  # allows to track ajax requests
